Refactored formatDateforInput to utils
